+++ custom_components/econet300/binary_sensor.py
@@ -68,7 +68,7 @@ BINARY_SENSOR_TYPES: tuple[EconetBinarySensorEntityDescription, ...] = (
     EconetBinarySensorEntityDescription(
         availability_key="pumpCO",
         key="pumpCOWorks",
-        name="pumpCO",
+        name="Pump",
         icon="mdi:pump",
         icon_off="mdi:pump-off",
         device_class=BinarySensorDeviceClass.RUNNING,
@@ -86,6 +86,7 @@ BINARY_SENSOR_TYPES: tuple[EconetBinarySensorEntityDescription, ...] = (
         key="feederWorks",
         name="Feeder",
         icon="mdi:screw-lag",
+        icon_off="mdi:screw-lag",
         device_class=BinarySensorDeviceClass.RUNNING,
     ),
     EconetBinarySensorEntityDescription(